+++ opre_ops/opre_ops/apps/hello_world/admin.py
