+++ opre_ops/opre_ops/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for opre_rpm project.
+ASGI config for opre_ops project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'opre_rpm.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'opre_ops.settings')
 
 application = get_asgi_application()
