+++ opre_ops/opre_ops/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for opre_rpm project.
+Django settings for opre_ops project.
 
 Generated by 'django-admin startproject' using Django 3.2.5.
 
@@ -49,7 +49,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'opre_rpm.urls'
+ROOT_URLCONF = 'opre_ops.urls'
 
 TEMPLATES = [
     {
@@ -67,7 +67,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'opre_rpm.wsgi.application'
+WSGI_APPLICATION = 'opre_ops.wsgi.application'
 
 
 # Database
