+++ opre_ops/opre_ops/urls.py
@@ -1,4 +1,4 @@
-"""opre_rpm URL Configuration
+"""opre_ops URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/3.2/topics/http/urls/
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from opre_rpm.apps.hello_world import views
+from opre_ops.apps.hello_world import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
