+++ opre_ops/opre_ops/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for opre_rpm project.
+WSGI config for opre_ops project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'opre_rpm.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'opre_ops.settings')
 
 application = get_wsgi_application()
