+++ start-web.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-python ./opre_rpm/manage.py runserver 0.0.0.0:8080
+python ./opre_ops/manage.py runserver 0.0.0.0:8080
